Add a content-addressed cache for wheels (#38)
Closes https://github.com/astral-sh/puffin/issues/4.

+++ crates/puffin-client/src/api.rs
@@ -147,7 +147,7 @@ pub struct File {
     pub filename: String,
     pub hashes: Hashes,
     pub requires_python: Option<String>,
-    pub size: i64,
+    pub size: usize,
     pub upload_time: String,
     pub url: String,
     pub yanked: Yanked,
